Update persyst.py
A bug where annotations with same descriptions were being overwritten by persyst annotation parser. In this fix annotations are added as lists of tuples to the repeated instances. Later when annotations objects are created repeated instances are added to the annotations objects. 

Of note, when transforming annotations to events it is not recognizing the repeated instances.